Extract MarketPriceService interface from DefaultMarketPriceService and use it instead of impl. class where possible
