+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/user/profile_card/offers/ProfileCardOffersController.java
@@ -18,8 +18,8 @@
 package bisq.desktop.main.content.user.profile_card.offers;
 
 import bisq.bisq_easy.BisqEasyOfferbookMessageService;
-import bisq.desktop.navigation.NavigationTarget;
 import bisq.bonded_roles.market_price.MarketPriceService;
+import bisq.desktop.navigation.NavigationTarget;
 import bisq.chat.bisq_easy.offerbook.BisqEasyOfferbookChannelService;
 import bisq.chat.bisq_easy.offerbook.BisqEasyOfferbookMessage;
 import bisq.chat.bisq_easy.offerbook.BisqEasyOfferbookSelectionService;
